Flyttet ut instansieringen av objekter i en egen fil.
Gjør dette for at `bootstrap.kt` skal være mest mulig ryddig, og at vi da følger samme pattern som for resten av Ktor-appene våre.
